+++ src/Vertr.Exchange.Terminal.ConsoleApp/Program.cs
@@ -1,6 +1,7 @@
 using Refit;
 using Vertr.Exchange.Terminal.ApiClient;
 using Vertr.Exchange.Terminal.ConsoleApp.StaticData;
+using Vertr.Exchange.Terminal.ConsoleApp.Views;
 
 namespace Vertr.Exchange.Terminal.ConsoleApp;
 
@@ -11,24 +12,27 @@ public class Program
         var exchApi = RestService.For<IHostApiClient>("http://localhost:5010");
 
         var res = await Commands.Reset(exchApi);
-        Console.WriteLine(res);
+        // Console.WriteLine(res);
 
         res = await Commands.AddSymbols(exchApi);
-        Console.WriteLine(res);
+        // Console.WriteLine(res);
 
         res = await Commands.AddUsers(exchApi);
-        Console.WriteLine(res);
+        // Console.WriteLine(res);
 
         var bobTrading = Task.Run(async () =>
         {
-            await TradingStrategy.RandomWalkTrading(exchApi, Users.Bob, Symbols.MSFT, 100m);
+            await TradingStrategy.RandomWalkTrading(exchApi, Users.Bob, Symbols.MSFT, 100m, 0.01m, 100);
         });
 
         var aliceTrading = Task.Run(async () =>
         {
-            await TradingStrategy.RandomWalkTrading(exchApi, Users.Alice, Symbols.MSFT, 100m);
+            await TradingStrategy.RandomWalkTrading(exchApi, Users.Alice, Symbols.MSFT, 100m, 0.01m, 100);
         });
 
         await Task.WhenAll(aliceTrading, bobTrading);
+
+        var ob = await exchApi.GetOrderBook(Symbols.MSFT.Id);
+        OrderBookView.Render(ob);
     }
 }