+++ src/Vertr.Exchange.Terminal.Server/BackgroundServices/OrderBooksService.cs
@@ -1,31 +1,30 @@
 using Microsoft.AspNetCore.SignalR.Client;
 using Vertr.Exchange.Contracts;
-using Vertr.Exchange.Terminal.Server.Awaiting;
 using Vertr.Exchange.Terminal.Server.Providers;
+using Vertr.Exchange.Terminal.Server.Repositories;
 
 namespace Vertr.Exchange.Terminal.Server.BackgroundServices;
 
-public class CommandResultService(
+public class OrderBooksService(
     HubConnectionProvider hubConnectionProvider,
-    ICommandAwaitingService commandAwaitingService,
+    IOrderBookRepository orderBookRepository,
     ILogger<CommandResultService> logger) : BackgroundService
 {
     private readonly HubConnectionProvider _connectionProvider = hubConnectionProvider;
+    private readonly IOrderBookRepository _orderBookRepository = orderBookRepository;
     private readonly ILogger<CommandResultService> _logger = logger;
-    private readonly ICommandAwaitingService _commandAwaitingService = commandAwaitingService;
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
-        _logger.LogInformation($"Start listening API Command result stream...");
+        _logger.LogInformation($"Start listening Order Books stream...");
         var connection = await _connectionProvider.GetConnection();
-        var channel = await connection.StreamAsChannelAsync<ApiCommandResult>("ApiCommandResults", stoppingToken);
+        var channel = await connection.StreamAsChannelAsync<OrderBook>("OrderBooks", stoppingToken);
         while (await channel.WaitToReadAsync(stoppingToken) && !stoppingToken.IsCancellationRequested)
         {
-            while (channel.TryRead(out var apiCommandResult))
+            while (channel.TryRead(out var orderBook))
             {
-                _logger.LogInformation("API Commad result received: {commandResult}", apiCommandResult);
-                var resp = new CommandResponse(apiCommandResult);
-                _commandAwaitingService.Complete(resp);
+                await _orderBookRepository.Save(orderBook);
+                _logger.LogInformation("Order Book received: {orderBook}", orderBook);
             }
         }
     }